+++ src/idunno.Bluesky/Notifications/Server/ListNotifications.cs
@@ -1,0 +1,122 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+using System.Text;
+
+using idunno.AtProto;
+using idunno.AtProto.Authentication;
+using idunno.Bluesky.Notifications;
+using idunno.Bluesky.Notifications.Model;
+
+using Microsoft.Extensions.Logging;
+
+namespace idunno.Bluesky;
+
+public static partial class BlueskyServer
+{
+    /// <summary>
+    /// Gets the notifications for the requesting account.
+    /// </summary>
+    /// <param name="limit">The maximum number of notifications to return. If specified this should be greater than 1 and less than or equal to 100.</param>
+    /// <param name="cursor">An optional cursor. See https://atproto.com/specs/xrpc#cursors-and-pagination.</param>
+    /// <param name="seenAt">The date and time notifications were last checked.</param>
+    /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
+    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+    /// <param name="subscribedLabelers">A optional list of labeler <see cref="Did"/>s to accept labels from.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when any of <paramref name="service"/>, <paramref name="accessCredentials"/> or <paramref name="httpClient"/> are <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is less than 1 or greater than 100.</exception>
+    [UnconditionalSuppressMessage(
+        "Trimming",
+        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    [UnconditionalSuppressMessage("AOT",
+        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    public static async Task<AtProtoHttpResult<NotificationCollection>> ListNotifications(
+        int? limit,
+        string? cursor,
+        DateTimeOffset? seenAt,
+        Uri service,
+        AccessCredentials accessCredentials,
+        HttpClient httpClient,
+        Action<AtProtoCredential>? onCredentialsUpdated = null,
+        ILoggerFactory? loggerFactory = default,
+        IEnumerable<Did>? subscribedLabelers = null,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(service);
+        ArgumentNullException.ThrowIfNull(accessCredentials);
+        ArgumentNullException.ThrowIfNull(httpClient);
+
+        if (limit is not null)
+        {
+            ArgumentOutOfRangeException.ThrowIfLessThan((int)limit, 1);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan((int)limit, 100);
+        }
+
+        StringBuilder queryString = new();
+        if (limit is not null)
+        {
+            queryString.Append(CultureInfo.InvariantCulture, $"limit={limit}&");
+        }
+
+        if (!string.IsNullOrEmpty(cursor))
+        {
+            queryString.Append(CultureInfo.InvariantCulture, $"cursor={Uri.EscapeDataString(cursor)}&");
+        }
+
+        if (seenAt is not null)
+        {
+            queryString.Append(CultureInfo.InvariantCulture, $"seenAt={seenAt.Value.UtcDateTime.ToString("o", CultureInfo.InvariantCulture)}");
+        }
+
+        if (queryString.Length > 0 && queryString[queryString.Length - 1] == '&')
+        {
+            queryString.Remove(queryString.Length - 1, 1);
+        }
+
+        BlueskyHttpClient<ListNotificationsResponse> request = new(AppViewProxy, loggerFactory);
+        AtProtoHttpResult<ListNotificationsResponse> response = await request.Get(
+            service: service,
+            endpoint: $"/xrpc/app.bsky.notification.listNotifications?{queryString}",
+            credentials: accessCredentials,
+            httpClient: httpClient,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
+            onCredentialsUpdated: onCredentialsUpdated,
+            subscribedLabelers: subscribedLabelers,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+
+        // Transform from DTO and flatten
+        if (response.Succeeded)
+        {
+            List<Notification> notifications = [];
+            foreach (NotificationResponse notificationResponse in response.Result.Notifications)
+            {
+                notifications.Add(new Notification(notificationResponse));
+            }
+
+            return new AtProtoHttpResult<NotificationCollection>(
+                new NotificationCollection(notifications, response.Result.Cursor, response.Result.Priority, response.Result.SeenAt),
+                response.StatusCode,
+                response.HttpResponseHeaders,
+                response.AtErrorDetail,
+                response.RateLimit);
+        }
+        else
+        {
+            return new AtProtoHttpResult<NotificationCollection>(
+                new(),
+                response.StatusCode,
+                response.HttpResponseHeaders,
+                response.AtErrorDetail,
+                response.RateLimit);
+        }
+    }
+}
