+++ src/idunno.Bluesky/PublicAPI.Shipped.txt
@@ -616,7 +616,7 @@ idunno.Bluesky.BlueskyAgent.Like(idunno.AtProto.AtUri! uri, idunno.AtProto.Cid! 
 idunno.Bluesky.BlueskyAgent.Like(idunno.AtProto.Repo.StrongReference! strongReference, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.AtProto.Repo.CreateRecordResult!>!>!
 idunno.Bluesky.BlueskyAgent.Like(idunno.Bluesky.Feed.FeedViewPost! post, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.AtProto.Repo.CreateRecordResult!>!>!
 idunno.Bluesky.BlueskyAgent.Like(idunno.Bluesky.Feed.PostView! postView, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.AtProto.Repo.CreateRecordResult!>!>!
-idunno.Bluesky.BlueskyAgent.ListActivitySubscriptions(int? limit = null, string? cursor = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.PagedViewReadOnlyCollection<idunno.Bluesky.Actor.ProfileView!>!>!>!
+idunno.Bluesky.BlueskyAgent.ListActivitySubscriptions(int? limit = null, string? cursor = null, System.Collections.Generic.IEnumerable<idunno.AtProto.Did!>? subscribedLabelers = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.PagedViewReadOnlyCollection<idunno.Bluesky.Actor.ProfileView!>!>!>!
 idunno.Bluesky.BlueskyAgent.ListConversationRequests(int? limit = null, string? cursor = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.PagedViewReadOnlyCollection<idunno.Bluesky.Chat.ConversationViewBase!>!>!>!
 idunno.Bluesky.BlueskyAgent.ListConversations(int? limit = null, string? cursor = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.Chat.Conversations!>!>!
 idunno.Bluesky.BlueskyAgent.ListJoinGroupRequests(System.Collections.Generic.ICollection<string!>! conversationIds, string? cursor = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.PagedViewReadOnlyCollection<idunno.Bluesky.Chat.Group.JoinRequestConversationView!>!>!>!
@@ -5399,6 +5399,7 @@ virtual Standard.Site.ThemeColorRgba.Equals(Standard.Site.ThemeColorRgba? other)
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetPostThreadOtherV2(idunno.AtProto.AtUri! anchor, System.Collections.Generic.IEnumerable<idunno.AtProto.Did!>? subscribedLabelers = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<System.Collections.Generic.IReadOnlyCollection<idunno.Bluesky.Unspecced.ThreadItem!>!>!>!
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetPostThreadV2(idunno.AtProto.AtUri! anchor, bool? above = true, int? below = 6, int? branchingFactor = 10, string? sort = "oldest", System.Collections.Generic.IEnumerable<idunno.AtProto.Did!>? subscribedLabelers = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.Unspecced.PostThreadV2!>!>!
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetSuggestedStarterPacks(int? limit = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<System.Collections.Generic.ICollection<idunno.Bluesky.Graph.StarterPackView!>!>!>!
+[BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetSuggestedUsers(string? category = null, int? limit = null, System.Collections.Generic.IEnumerable<idunno.AtProto.Did!>? subscribedLabelers = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<System.Collections.Generic.ICollection<idunno.Bluesky.Actor.ProfileView!>!>!>!
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetSuggestedUsers(string? category = null, int? limit = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<System.Collections.Generic.ICollection<idunno.Bluesky.Actor.ProfileView!>!>!>!
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetTaggedSuggestions(System.Collections.Generic.ICollection<System.Collections.Generic.KeyValuePair<string!, object!>>? parameters = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<System.Collections.Generic.ICollection<idunno.Bluesky.Unspecced.Suggestion!>!>!>!
 [BSKYUnspecced]idunno.Bluesky.BlueskyAgent.GetTrendingTopics(int? limit = null, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<idunno.AtProto.AtProtoHttpResult<idunno.Bluesky.Unspecced.TrendingTopics!>!>!
