+++ src/idunno.Bluesky/Unspecced/Agent/GetSuggestedUsers.cs
@@ -1,0 +1,51 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+
+using idunno.AtProto;
+using idunno.Bluesky.Actor;
+using idunno.Bluesky.Feed;
+using idunno.Bluesky.Graph;
+using idunno.Bluesky.Unspecced;
+
+namespace idunno.Bluesky;
+
+public partial class BlueskyAgent
+{
+    /// <summary>
+    /// Get a <see cref="PagedReadOnlyCollection{T}"/> of <see cref="ProfileView"/>s of suggested actors.
+    /// </summary>
+    /// <param name="category">An optional category of users to get suggestions for.</param>
+    /// <param name="limit">The number of topics to return. Must be between 1 and 50.</param>
+    /// <param name="subscribedLabelers">An optional list of <see cref="Did"/>s of labelers to retrieve labels applied to the account.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt;50.</exception>
+    [Experimental("BSKYUnspecced", UrlFormat = "https://bluesky.idunno.dev/docs/unspecced.html")]
+    public async Task<AtProtoHttpResult<ICollection<ProfileView>>> GetSuggestedUsers(
+        string? category = null,
+        int? limit = null,
+        IEnumerable<Did>? subscribedLabelers = null,
+        CancellationToken cancellationToken = default)
+    {
+        if (limit is not null)
+        {
+            ArgumentOutOfRangeException.ThrowIfLessThan(limit.Value, 1);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan(limit.Value, Maximum.SuggestedUsers);
+        }
+
+#pragma warning disable BSKYUnspecced
+        return await BlueskyServer.GetSuggestedUsers(
+            category,
+            limit,
+            service: Service,
+            accessCredentials: Credentials,
+            httpClient: HttpClient,
+            onCredentialsUpdated: InternalOnCredentialsUpdatedCallBack,
+            loggerFactory: LoggerFactory,
+            subscribedLabelers: subscribedLabelers,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+#pragma warning restore BSKYUnspecced
+    }
+}
